+++ packages/@react-native-windows/tester/src/js/utils/RNTesterList.windows.js
@@ -12,10 +12,13 @@ import type {RNTesterModuleInfo} from '../types/RNTesterTypes';
 import ReactNativeFeatureFlags from 'react-native/Libraries/ReactNative/ReactNativeFeatureFlags';
 
 const Components: Array<RNTesterModuleInfo> = [
-  {
+  /*{
+    key: 'DrawerLayoutAndroid',
+    category: 'UI',
+    module: require('../examples/DrawerLayoutAndroid/DrawerLayoutAndroidExample'),
     key: 'HTTPExample',
     module: require('../examples/HTTP/HTTPExample'),
-  },
+  },*/
   {
     key: 'XHRExample',
     module: require('../examples/XHR/XHRExample'),
@@ -34,7 +37,6 @@ const Components: Array<RNTesterModuleInfo> = [
     key: 'FlatListExampleIndex',
     module: require('../examples/FlatList/FlatListExampleIndex').default,
     category: 'ListView',
-    supportsTVOS: true,
   },
   {
     key: 'ImageExample',
@@ -354,6 +356,11 @@ const APIs: Array<RNTesterModuleInfo> = [
     key: 'TurboModuleExample',
     category: 'Basic',
     module: require('../examples/TurboModule/TurboModuleExample'),
+  },
+  {
+    key: 'TurboCxxModuleExample',
+    category: 'Basic',
+    module: require('../examples/TurboModule/TurboCxxModuleExample'),
   },
 ];