+++ packages/@react-native-windows/automation-channel/CHANGELOG.md
@@ -1,8 +1,40 @@
 # Change Log - @react-native-windows/automation-channel
 
-This log was last generated on Wed, 16 Nov 2022 06:16:01 GMT and should not be manually modified.
+This log was last generated on Thu, 24 Nov 2022 06:15:42 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.5.8
+
+Thu, 24 Nov 2022 06:15:42 GMT
+
+### Patches
+
+- Bump react-native-windows to v0.0.0-canary.587
+
+## 0.5.7
+
+Wed, 23 Nov 2022 06:16:32 GMT
+
+### Patches
+
+- Bump react-native-windows to v0.0.0-canary.586
+
+## 0.5.6
+
+Tue, 22 Nov 2022 06:15:33 GMT
+
+### Patches
+
+- Bump react-native-windows to v0.0.0-canary.585
+
+## 0.5.5
+
+Mon, 21 Nov 2022 06:15:36 GMT
+
+### Patches
+
+- Bump react-native-windows to v0.0.0-canary.584
 
 ## 0.5.4