+++ packages/react-native-windows-init/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Change Log - react-native-windows-init
 
-This log was last generated on Wed, 16 Nov 2022 06:16:01 GMT and should not be manually modified.
+This log was last generated on Wed, 23 Nov 2022 06:16:32 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 1.3.5
+
+Wed, 23 Nov 2022 06:16:32 GMT
+
+### Patches
+
+- Bump @react-native-windows/cli to v0.0.0-canary.152
+
+## 1.3.4
+
+Mon, 21 Nov 2022 06:15:36 GMT
+
+### Patches
+
+- Bump @react-native-windows/cli to v0.0.0-canary.151
 
 ## 1.3.3